cdk/requirements.txt
pytest==6.2.5
aws-cdk-lib==2.193.0
constructs>=10.0.0,<11.0.0
boto3
